refactor: hook and variables name
